If a program wasn't uploaded to VRAM yet, it won't have a valid handle, so handle that to avoid GL errors

+++ Source Code/Platform/Video/RenderBackend/OpenGL/Shaders/glShader.cpp
@@ -71,8 +71,10 @@ glShader::glShader(GFXDevice& context,
 }
 
 glShader::~glShader() {
-    Console::d_printfn(Locale::get(_ID("SHADER_DELETE")), name().c_str());
-    GL_API::deleteShaderPrograms(1, &_programHandle);
+    if (_programHandle != GLUtil::_invalidObjectID) {
+        Console::d_printfn(Locale::get(_ID("SHADER_DELETE")), name().c_str());
+        GL_API::deleteShaderPrograms(1, &_programHandle);
+    }
 }
 
 bool glShader::embedsType(ShaderType type) const {
